Huge update. Testing...

+++ summon.sh
@@ -6,11 +6,15 @@
 
 ssd=$(</note/ssd.txt)
 efi=$(</note/efi.txt)
+hostname=$(</note/hostname.txt)
 rootpwd=$(</note/rootpwd.txt)
-userpwd=$(</note/userpwd.txt)
 username=$(</note/username.txt)
+userpwd=$(</note/userpwd.txt)
+superuser=$(</note/superuser.txt)
 
 set -o pipefail
+
+reflector --latest 5 --sort rate --save /etc/pacman.d/mirrorlist
 
 timedatectl set-ntp true
 
@@ -28,7 +32,12 @@ timedatectl set-timezone Europe/Belgrade
 systemctl enable --now NetworkManager
 systemctl enable --now nftables.service
 systemctl enable --now paccache.timer
-systemctl enable lightdm
+if [ $superuser == 1 ]
+then
+        goo=1
+else
+        systemctl enable lightdm
+fi
 
 ######################################
 # CREATING USER & CHANGING PASSWORDS #
@@ -50,15 +59,28 @@ cp /Atina/files/sudoers /etc/
 rm -rf /etc/locale.gen
 cp /Atina/files/locale.gen /etc/
 locale-gen
-cp /Atina/files/locale.conf /etc/
+localeconf="LANG=en_US.UTF-8"
+$localeconf > /etc/locale.conf
 
 ############################################
 # SETTING UP HOSTNAME AND NETWORK SETTINGS #
 ############################################
 
-cp /Atina/files/hostname /etc/
-rm -rf /etc/hosts
-cp /Atina/files/hosts /etc/
+$hostname > /etc/hostname
+$base_hostname="# Static table lookup for hostnames.
+# See hosts(5) for details.
+
+127.0.0.1	localhost
+::1		localhost
+127.0.1.1	"
+plus="" ; plus="$hostname"
+ext=".localdomain $hostname"
+$base_hostname="$plus$ext"
+$base_hostname > /etc/hosts
+
+#cp /Atina/files/hostname /etc/
+#rm -rf /etc/hosts
+#cp /Atina/files/hosts /etc/
 
 ###########################
 # FINALIZING INSTALLATION #
@@ -72,14 +94,31 @@ rm -rf /home/$username/.bashrc
 cp /Atina/files/bashrc /home/$username/
 mv /home/$username/bashrc /home/$username/.bashrc
 
-cp /Atina/files/xinitrc /home/$username/
-mv /home/$username/xinitrc /home/$username/.xinitrc
+if [ $superuser == 1 ]
+then
+        cp /Atina/files/xinitrc-superuser /home/$username/
+        mv /home/$username/xinitrc-superuser /home/$username/.xinitrc
+else
+        cp /Atina/files/xinitrc /home/$username/
+        mv /home/$username/xinitrc /home/$username/.xinitrc
+fi
+
+#cp /Atina/files/xinitrc /home/$username/
+#mv /home/$username/xinitrc /home/$username/.xinitrc
 
 rm -rf /etc/pacman.conf
 cp /Atina/files/pacman.conf /etc/
 
-rm -rf /etc/lightdm/lightdm.conf
-cp /Atina/files/lightdm.conf /etc/lightdm/
+if [ $superuser == 1 ]
+then
+        goo=1
+else
+        rm -rf /etc/lightdm/lightdm.conf
+        cp /Atina/files/lightdm.conf /etc/lightdm/
+fi
+
+#rm -rf /etc/lightdm/lightdm.conf
+#cp /Atina/files/lightdm.conf /etc/lightdm/
 
 pacman -Sy
 bash /Atina/scrollbook/32bit.sh
@@ -104,12 +143,46 @@ fi
 sudo amixer sset "Auto-Mute Mode" Disabled
 sudo alsactl store
 
-#sudo pacman -Rns xfce4-terminal --noconfirm
 sudo pacman -Rns lxqt-archiver pcmanfm-qt qterminal lxqt-about --noconfirm # lrf-viewer lximage screenshot screengrab
 sudo pacman -Scc --noconfirm
 
 cd /home/$username/ ; mkdir .atina ; chmod ugo+rwx /home/$username/.atina/
 cp -r /Atina/scrollbook/ /home/$username/.atina/
 
+if [ $superuser == 1 ]
+then
+        cd /home/$username/ ; mkdir .myconfig
+        cd /home/$username/.myconfig
+        git clone https://aur.archlinux.org/yay.git
+        cd yay
+        makepkg -sic --noconfirm
+
+        cd ~/
+        git clone https://github.com/windwalk-bushido/AIFAL.git
+        git clone https://github.com/windwalk-bushido/Atina.git
+
+        cp /home/$username/AIFAL/scrollbook/search.sh ~/.atina/scrollbook/
+        cp /home/$username/AIFAL/scrollbook/yay.sh ~/.atina/scrollbook/
+        cp /home/$username/AIFAL/scrollbook/mount-dvd.sh ~/.atina/scrollbook/
+        cp /home/$username/AIFAL/scrollbook/unmount-dvd.sh ~/.atina/scrollbook/
+        cd /home/$username/
+
+        chmod ugo+rwx /home/$username/.atina/scrollbook/
+
+        rm -rf /home/$username/.bashrc
+        cp /home/$username/Atina/files/bashrc-superuser /home/$username/
+        mv /home/$username/bashrc-superuser /home/$username/.bashrc
+
+        cd /home/$username/ ; rm -rf /home/$username/AIFAL ; rm -rf /home/$username/Atina
+
+        yay -S ttf-iosevka gedit-latex cherrytree labyrinth peaclock --noconfirm
+
+        yay -Scc --noconfirm
+        sudo pacman -Scc --noconfirm
+else
+        goo=1
+fi
+
+cp /note/superuser.txt /home/$username/
 rm -rf /note/
 rm -rf /Atina/
